# 3. Admin with Email Login (Includes Super Admin & Customer)/server/swagger.yaml
openapi: 3.0.0
paths:
...
  /xam/email-templates:
    get:
      operationId: EmailTemplateController_getAllTemplates
      parameters: []
      responses:
        '200':
          description: List of email templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmailTemplateDto'
      summary: Get all email templates
      tags: &ref_2
        - Email Templates
    post:
      operationId: EmailTemplateController_createEmailTemplate
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmailTemplateDto'
      responses:
        '201':
          description: Email template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTemplateDto'
      summary: Create a new email template
      tags: *ref_2
  /xam/email-templates/{id}:
    get:
      operationId: EmailTemplateController_getTemplateById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Email template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTemplateDto'
      summary: Get email template by ID
      tags: *ref_2
    put:
      operationId: EmailTemplateController_updateTemplate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmailTemplateDto'
      responses:
        '200':
          description: Email template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTemplateDto'
      summary: Update an existing email template
      tags: *ref_2
    delete:
      operationId: EmailTemplateController_deleteTemplate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Email template deleted
      summary: Delete an email template
      tags: *ref_2
...
info:
  title: THIS BACKEND MADE BY SUBHAM MAITY
  description: THIS APP IS MADE FOR EVERYONE
  version: 1.0.0
  contact:
    name: SUBHAM MAITY
    url: https://github.com/Subham-Maity
    email: maitysubham4041@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
servers:
  - url: http://localhost:3333
    description: Local server (development)
  - url: http://210.89.44.183:3333/xam
    description: Staging server (testing)
  - url: http://210.89.44.183:3333/xam
    description: Production server (live)
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    EmailTemplateDto:
      type: object
      properties:
        id:
          type: string
          description: ID of the email template
          example: clabc123
        name:
          type: string
          description: Unique name of the email template
          example: OrderConfirmation
        subject:
          type: string
          description: Subject of the email
          example: Your Order Confirmation
        html:
          type: string
          description: HTML content with placeholders
          example: <p>Hello {{userName}}, your order {{orderId}} is confirmed.</p>
        createdAt:
          format: date-time
          type: string
          description: Creation timestamp
          example: '2023-01-01T00:00:00Z'
        updatedAt:
          format: date-time
          type: string
          description: Last update timestamp
          example: '2023-01-01T00:00:00Z'
      required:
        - id
        - name
        - subject
        - html
        - createdAt
        - updatedAt
    CreateEmailTemplateDto:
      type: object
      properties:
        name:
          type: string
          description: Unique name of the email template
          example: OrderConfirmation
        subject:
          type: string
          description: Subject of the email
          example: Your Order Confirmation
        html:
          type: string
          description: HTML content with placeholders
          example: <p>Hello {{userName}}, your order {{orderId}} is confirmed.</p>
      required:
        - name
        - subject
        - html
    UpdateEmailTemplateDto:
      type: object
      properties:
        name:
          type: string
          description: Unique name of the email template
          example: OrderConfirmation
        subject:
          type: string
          description: Subject of the email
          example: Your Order Confirmation
        html:
          type: string
          description: HTML content with placeholders
          example: <p>Hello {{userName}}, your order {{orderId}} is confirmed.</p>
...